feat: package_cpp completion バグあるかも　とりあえず置換した
